+++ src/IronyModManager.Services/ModService.cs
@@ -4,7 +4,7 @@
 // Created          : 02-24-2020
 //
 // Last Modified By : Mario
-// Last Modified On : 06-23-2020
+// Last Modified On : 07-11-2020
 // ***********************************************************************
 // <copyright file="ModService.cs" company="Mario">
 //     Mario
@@ -424,10 +424,10 @@ namespace IronyModManager.Services
         /// <param name="path">The path.</param>
         /// <param name="isDirectory">if set to <c>true</c> [is directory].</param>
         /// <returns>System.Int32.</returns>
-        protected virtual int GetSteamModId(string path, bool isDirectory = false)
+        protected virtual long GetSteamModId(string path, bool isDirectory = false)
         {
             var name = !isDirectory ? Path.GetFileNameWithoutExtension(path) : path;
-            int.TryParse(name.Replace(Constants.Steam_mod_id, string.Empty), out var id);
+            long.TryParse(name.Replace(Constants.Steam_mod_id, string.Empty), out var id);
             return id;
         }
